requirements.txt
requests==2.32.5
python-telegram-bot==22.5